backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic[email]==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
aiofiles==23.2.1
tenacity==8.2.3
pyyaml==6.0.1
requests==2.31.0
